+++ docs/md/releasenotes.md
@@ -18,6 +18,7 @@
 - Fix wrong offset for complementary branch instruction when widening branch instructions in `InstructionWriter`. (`T5721`)
 - Fix potential `ClassFormatError` due to adding multiple annotation attributes when processing Kotlin code.
 - Fix potential `NullPointerException` due to missing classes in `ClassReferenceInitializer`.
+- Prevent making package-private final methods that are shadowed protected. (`T7056`)
 
 ## Version 8.0.1